Changed the tests for login and accessing protected resources
SiteMinder does not deal with login validation by redirecting the client until a request is made for a protected resource. This means that POST requests to the login page must create state in the form of a FORMCRED cookie that tracks the outcome of the login request. When a request is made for a protected resource with a FORMCRED cookie then the cookie is read and the appropriate action taken.