Fixed bug that allowed locked users to access protected URLs
Save method for User model objects
Save updates/add the current User to the specified object collection or array.
Fixed a bug to ensure locked status for a user is persisted correctly
Reset login_attempts back to zero after a successful login. Logon successfully before submitting bad passwords to ensure that the user's login_attempts is set back to zero. This ensures an accurate count of logon attempts and that no other tests affect this one.