Support for more advanced URL pattern matching definitions
Wildcards are now supported.
Each URL pattern can be marked as protected or not.
A default protection level can now be configured if a matching pattern is not found for a given URL.
Also moved the path definitions for redirects to the siteminder section of the configuration. This makes more sense currently as the redirect are a SiteMinder specific configuration.